Create scrollable html with css
In the example above, I have created a `<div>` element with the class "scrolling-text-container" that represents the container for the scrolling text. I have set the dimensions of the container to a width of 300px and a height of 200px, but you can customize them as desired.

Inside the container, I have created another `<div>` element with the class "scrolling-text" that contains the text you want to scroll. I have applied a CSS animation called "scroll" to this element, which horizontally translates the text inside the container for its entire width over a period of 10 seconds.

Modify the text inside the `<div class="scrolling-text">` element according to your needs. You can also customize the CSS style of the "scrolling-text-container" and "scrolling-text" classes to match the design of your website.